Update tests
- move canvas tests out of text_truncate

- update existing method of mocking canvas width, since the new `testenv` file now allows bypassing setting `HTMLCanvasElement`
